refactor: replace speed-highlight with rangi
Swap @speed-highlight/core for pi0/rangi as the lightweight highlighter.
- comark/plugins/rangi with theme pairs, inline colors, optional line wrappers
- docs, Vue+Vite example, tests, peer dep (rangi)
- remove speed-highlight plugin, docs, and example
