+++ examples/3.plugins/vue-vite-rangi/index.html
@@ -2,11 +2,8 @@
 <html lang="en">
   <head>
     <meta charset="UTF-8" />
-    <meta
-      name="viewport"
-      content="width=device-width, initial-scale=1.0"
-    />
-    <title>Comark + Speed Highlight - Vue Example</title>
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>Comark + Rangi - Vue Example</title>
     <style>
       body {
         margin: 0;
@@ -80,9 +77,6 @@
   </head>
   <body>
     <div id="app"></div>
-    <script
-      type="module"
-      src="/src/main.ts"
-    ></script>
+    <script type="module" src="/src/main.ts"></script>
   </body>
 </html>
