+++ examples/3.plugins/vue-vite-rangi/src/main.ts
