+++ examples/3.plugins/vue-vite-rangi/tsconfig.json
