+++ examples/3.plugins/vue-vite-rangi/vite.config.ts
