+++ package.json
@@ -27,7 +27,7 @@
     "dev:sveltekit": "pnpm --filter comark-sveltekit run dev",
     "dev:nextjs": "pnpm --filter comark-nextjs run dev",
     "dev:highlight": "pnpm --filter comark-vue-vite-highlight run dev",
-    "dev:speed-highlight": "pnpm --filter comark-vue-vite-speed-highlight run dev",
+    "dev:rangi": "pnpm --filter comark-vue-vite-rangi run dev",
     "dev:twoslash": "pnpm --filter comark-vue-vite-twoslash run dev",
     "dev:json-render": "pnpm --filter comark-vue-vite-json-render run dev",
     "dev:binding": "pnpm --filter comark-vue-vite-binding run dev",
