+++ packages/comark/src/internal/stringify/attributes.ts
@@ -101,7 +101,7 @@ const IMPLICIT_ATTRS: Record<string, { drop?: string[]; classBlocklist?: string[
   ul: { classBlocklist: ['contains-task-list'] },
   li: { classBlocklist: ['task-list-item'] },
   // `language`/`filename`/`highlights`/`meta` ride on the fence info string.
-  // `style` comes from render-time plugins (e.g. shiki / speed-highlight) and has no
+  // `style` comes from render-time plugins (e.g. shiki / rangi) and has no
   // markdown form. `class` is handled specially in userBlockAttrs because
   // highlighters merge their injected classes with the user's class — we need
   // to strip just the highlighter portion.
