+++ pnpm-workspace.yaml
@@ -120,7 +120,7 @@ catalog:
   scule: ^1.3.0
 
   # Shared peer dependencies
-  "@speed-highlight/core": ^1.2.23
+  rangi: ^2.2.0
   shiki: ^4.3.1
   splitpanes: ^4.0.4
 
@@ -173,6 +173,7 @@ minimumReleaseAgeExclude:
   - comark@0.5.1 || 0.6.0
   - "@comark/vue@0.5.1 || 0.6.0"
   - "@nuxtlabs/monarch-mdc@0.10.0"
+  - rangi@2.2.0
 
 patchedDependencies:
   minimark: patches/minimark.patch
